# Conflicts:
#	.gitmodules
#	CMakeLists.txt
#	lib/imgui
#	src/Cool/AppManager/AppManager.cpp
#	src/Cool/AppManager/AppManager.h
#	src/Cool/Core/run.cpp
#	src/Cool/Gpu/OpenGL/FullscreenPipeline.cpp
#	src/Cool/Gpu/OpenGL/FullscreenPipeline.h
#	src/Cool/Gpu/OpenGL/Shader.cpp
#	src/Cool/Gpu/OpenGL/TextureFB.cpp
#	src/Cool/Gpu/Shader.h
#	src/Cool/Gpu/TextureSource_FromFile.cpp
#	src/Cool/TextureSource/TextureDescriptor.h
#	src/Cool/TextureSource/TextureLibrary_Image.cpp
#	src/Cool/TextureSource/TextureLibrary_Image.h
#	src/Cool/TextureSource/TextureSamplerLibrary.cpp
#	src/Cool/TextureSource/TextureSamplerLibrary.h
#	src/Cool/TextureSource/TextureSource.cpp
#	src/Cool/TextureSource/TextureSource.h
#	src/Cool/View/View.cpp
#	src/Cool/Webcam/internal/WebcamCapture.cpp

+++ src/Cool/Exporter/internal/ImageExportJob.h
@@ -8,7 +8,7 @@ namespace Cool {
 class ImageExportJob {
 public:
     ImageExportJob() = default;
-    ImageExportJob(std::filesystem::path file_path, img::Image&& image, Averager<float>& average_export_time, std::mutex& average_export_time_mutex, std::atomic<int>& nb_frames_which_finished_exporting)
+    ImageExportJob(std::filesystem::path file_path, img::Image&& image, Averager<double>& average_export_time, std::mutex& average_export_time_mutex, std::atomic<int64_t>& nb_frames_which_finished_exporting)
         : _file_path{file_path}
         , _image{std::move(image)}
         , _average_export_time{&average_export_time}
@@ -24,9 +24,9 @@ public:
 private:
     std::filesystem::path     _file_path;
     std::optional<img::Image> _image;
-    Averager<float>*          _average_export_time;
+    Averager<double>*         _average_export_time;
     std::mutex*               _average_export_time_mutex;
-    std::atomic<int>*         _nb_frames_which_finished_exporting;
+    std::atomic<int64_t>*     _nb_frames_which_finished_exporting;
 };
 
 } // namespace Cool
